+++ lib/service_layer.rb
@@ -7,4 +7,5 @@ require 'service_layer/version'
 module ServiceLayer
   autoload :Base, 'service_layer/base'
   autoload :Command, 'service_layer/command'
+  autoload :Contract, 'service_layer/contract'
 end