+++ spec/service_layer/contract_spec.rb
@@ -18,4 +18,21 @@ RSpec.describe ServiceLayer::Contract do
       is_expected.to have_attributes(properties)
     end
   end
+
+  describe '#render' do
+    subject { contract.render }
+
+    let(:contract) { described_module.new }
+
+    before { described_module.render :domain }
+
+    it 'returns a Result' do
+      is_expected.to be_a(ServiceLayer::Result)
+    end
+
+    it 'returns a Result containing the rendering attributes' do
+      contract.domain = 'gmail.com'
+      is_expected.to have_attributes(domain: 'gmail.com')
+    end
+  end
 end