+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add the `Contract` structure. The Contract design pattern allows to setup
 properties of a `Class` and auto-initialize this ones.
+* Allow `Contract` to define, with `.render`, the rendered properties. These
+will be auto-transferred to the `Result` returned with `#render`.
 
 ## [0.3.0] (2017-12-28)